Revert "Merge branch 'master' into Shadow-combining-and-flask-unlocking"
This reverts commit 59177c775069661df8379f672e4a7f0bc3841a24, reversing
changes made to 68aa8485817d483824c3e7a9d333378e1fecbdfe.
