--- Assets/Art/Animations/Creatures.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f22b15225d3d04338a450b4eecc67296
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
