--- Assets/Art/Animations/Creatures/CottonSpiteSit.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 49e491b7005b44740a8a5fc4ecb62bc1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
