--- Assets/Art/Animations/Creatures/CottonSprite.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 968159b1e4ae24f1a8f2361ae58d0a6e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
