--- Assets/Art/Animations/Creatures/CottonSpriteStand.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: adcba0dca30d9465581bc93c337c5bcb
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
