--- Assets/Art/Animations/Creatures/CottonSpriteWalk.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6f1f240804a2947678d3395f1dbfb65e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
