--- Assets/Art/Animations/Creatures/Flittlet.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2be86a0c5fa3e48f7bb2fed9ae1e83ee
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
