--- Assets/Art/Animations/Creatures/Flittlet.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 696e549ec554f4046b6fbb2a5020bfe7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
