--- Assets/Art/Music.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e5f552c0ada694ea9a01c7681e451a8a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
