--- Assets/Prefabs/CottonSprite.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7a6976e5e10a12443bc9531d0c111a7b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
