+++ Assets/Prefabs/Environment/Tree1b.prefab
@@ -101,7 +101,7 @@ MeshCollider:
     m_Bits: 0
   m_ExcludeLayers:
     serializedVersion: 2
-    m_Bits: 64
+    m_Bits: 0
   m_LayerOverridePriority: 0
   m_IsTrigger: 0
   m_ProvidesContacts: 0
@@ -122,7 +122,7 @@ NavMeshObstacle:
   m_Shape: 1
   m_Extents: {x: 0.014, y: 0.016, z: 0.032}
   m_MoveThreshold: 0.1
-  m_Carve: 1
+  m_Carve: 0
   m_CarveOnlyStationary: 1
   m_Center: {x: 0, y: -0.002, z: 0.032}
   m_TimeToStationary: 0.5
