--- Assets/Prefabs/Flittlet.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d5b61004eef7bc04aaeea03c6c550b10
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
