--- Assets/Scripts/Creatures/Shadows/Flittlet.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5ebab339b57964c6fa48edb92fe774c4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
