+++ Assets/Scripts/Creatures/StateMachine.cs
@@ -20,6 +20,7 @@ public class StateMachine : MonoBehaviour
         if (activeState != null)
         {
             activeState.Perform();
+            Debug.Log("Active State: " + activeState);
         }
     }
