--- Assets/Scripts/Creatures/States/CottonSprite.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 602d4b1b2d4301248bcc83f83745bc92
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
