--- Assets/Scripts/Creatures/States/CottonSprite/CottonSpriteFollowState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e737d9b958d28cf4580093739ba6456d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
