--- Assets/Scripts/Creatures/States/CottonSprite/CottonSpriteWanderState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c843957157e913342882d394db67526b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
