--- Assets/Scripts/Creatures/States/Flittlet.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 40086423bc708304d967a72bff037072
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
