--- Assets/Scripts/Creatures/States/Flittlet/FlittletWanderState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b5d9ce00f32857a488fc38d3bc344166
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
