+++ Assets/Scripts/Creatures/States/FollowState.cs
@@ -1,0 +1,27 @@
+using System.Collections;
+using System.Collections.Generic;
+using System.Text;
+using UnityEngine;
+
+public class FollowState : BaseState
+{
+    public FollowState(StateMachine stateMachine) : base(stateMachine) {}
+    public override void Enter()
+    {
+
+    }
+
+    public override void Perform()
+    {
+
+    }
+    public override void Exit()
+    {
+
+    }
+
+    public void FollowPlayer()
+    {
+
+    }
+}
