+++ Assets/Scripts/Creatures/States/SittingState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64d18bc4680765b47978181ca8497f54
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
