+++ Assets/Scripts/Creatures/States/WanderState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c843957157e913342882d394db67526b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
