--- Assets/Scripts/Interactables/CottonSprite.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d7bb0274e4c5d489e8415d0133c7b0a4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
