+++ ProjectSettings/TagManager.asset
@@ -11,7 +11,7 @@ TagManager:
   - PostProcessing
   - Water
   - UI
-  - Creatures
+  - 
   - 
   - 
   -
